Started restructuring. Crawler almost obsolete.
